src/base/test2/gamma_testing/requirements.txt
grpcio
grpcio-tools
opencv-python
numpy

